Fix for the fix of Blueprint validations.
Discussed with @mahagr.

Saved YAML configs gave undesired extra parameter (!!float 1 for example) because floats were not being cast back to integers upon save. This was even true when the filterNumber function was giving back a correct parameter. filter_var from validateFloat was actually perserving the float variable type which the YAML engine perserved upon Yaml::dump.

This is a unconventional fix, but it is the simplest way to handle this edge case.