Fixed what I think may have been an oversight.
getRawContent and setRawContent were structured to get and set the content AFTER the content had already been processed. I ran into this problem while developing a plugin for my own personal use. If this was not an oversight and was in fact the intended behavior, then the wording is very misleading. lol.