Avoid to delete images/ folder after using GPM
When I installed or updated a Grav plugin, it deleted all the cache, but also the "compiled" images,
makes the website very slow (for a relatively complex website).

My initial reason for this patch is because I installed fastcgi_cache in order to make the site still available while
deploying with Ansible, but the images got cleared when the provisioning run the update GPM command,
then showing "missing picture" errors on the website.

This patch nows permit to use the function like :

```php
Grav\Console\ConsoleTrait::clearCache('all');
Grav\Console\ConsoleTrait::clearCache('cache-only');
```

*Please not that those examples are static, just consider the method parameter.*

And then I've linked it after the update and install classes. So when a plugin got updated/installed,
it does only clear the cache instead of clear the cache plus the compiled pictures, agressively slowing the website.

In hope that this will break nothing in the actual core, because I don't know Grav well enough to deal with all the parameters in my mind.

Thanks !