Fix taxonomy sort
Assume you have the following structure:

/home
/home/featured-work
/portfolio
/portfolio/portfolio-item-1
/portfolio/portfolio-item-2
/portfolio/portfolio-item-3

/home is a modular page, and the /home/featured-work page is a module that
displays a number of pages with a given taxonomy from the /portfolio collection.

If each portfolio item is manually sorted (i.e. by folder prefix), then viewing
the /portfolio page displays the collection in the *correct* order.

However, when items from this collection are rendered by a page under a sibling
parent, the sort order is *incorrect*.

I believe this is a bug, but it would be good to confirm.

Looking at the code, it appears to be a straightforward fix, sorting the
collection generated from a given taxonomy.

Test-information:

By cp'ing the files over an existing Grav site and testing that items displayed
by a module on the home page (from a sibling collection) were displayed as
expected.