Changing back the default redirect code from 301 to 302.
A permanent redirect can not be changed remotely once it has been sent.

Permanent redirects (301 / 308) are exactly what the their descriptive
name suggest. A 301 or 308 will be indefinitely cached client-side,
sometimes even when a user clears their cache.

They should be used to keep search engine page rankings by sending a
permanent redirect on a page by page basis, not for site to site or
protocols (http to https).

Especially on a new Grav installation, the default should not be 301.

My reasoning is that with the first visit to a new installation, before
even doing anything else, the 301 is already permanently cached in the
browser.

Additionally, temporary redirects can be influenced remotely and are not
that expensive on the webserver nowadays with event-based handlers.