Do not remove trailing spaces on markdown files
Using Jetbrains IDEA IDE like PHPStorm when you have the setting "strip trailing spaces on save" selected (which is useful for all but markdown files) it will always remove 3 spaces added deliberately to force a line break and mess up the output to have all on one line.

See also https://github.com/nicoulaj/idea-markdown/issues/138 (the linked plugin is outdated and replaces by vendor markdown plugin but the `.editorconfig` fix described and proposed with this PR fixes the problem for me. Not sure though as nobody seems to have that issue here. Afaik the mentioned setting causing it is disabled by default, so this could be why it rarely happens.