Correct width and height after image manipulation
I have various size images (both horizontal and vertical). All are at least 1900px on longest side (some 2000px)
```twig
{% set image = item.cropResize(1600, 1600) %}
{{ image.width ~ 'x' ~ image.height }}
```
Result is something like `2000x1350` or `1200x1900`, but it should be 1600 on longest side

Debugged a bit and found the solution, but was afraid, that calling ` $this->image->get()` to actually save the image to get correct dimensions might not be the way. Searched a bit and found [this issue on Gregwar/Image](https://github.com/Gregwar/Image/issues/123) and there's [a fix and a comment](https://github.com/Gregwar/Image/issues/123#ref-commit-948240d) from Gregwar, that this is the way.

This change will cache image with default quality from config and return correct cropped/resized dimensions